Consolidate command improvements
- Improved suggested filename generation for the Consolidate command. No longer suggests a filename that already exists.
- The Consolidate command now asks user whether they want to overwrite a file that already exists.
- Updated readme accordingly

+++ client/out/commands/epilog_consolidate.js
@@ -5,13 +5,30 @@ const vscode = require("vscode");
 const fs = require("fs");
 const resolve_full_file_content_js_1 = require("../../../common/out/resolve_full_file_content.js");
 const path = require("path");
+function generateUnusedFilename(extension, currentDirectory, prefix) {
+    let filename = (prefix ?? 'newfile') + extension;
+    if (!fs.existsSync(currentDirectory + '/' + filename)) {
+        return filename;
+    }
+    // If the filename already exists, add a number to the end of the filename until it doesn't exist
+    let i = 1;
+    while (fs.existsSync(currentDirectory + '/' + filename)) {
+        filename = prefix + i + extension;
+        i++;
+    }
+    return filename;
+}
 async function epilogCmd_consolidate(client) {
     // Get the uri of the active document
     let documentAbsFilepath = vscode.window.activeTextEditor.document.uri.fsPath;
+    // Get the directory of the active document
+    let documentDir = path.dirname(documentAbsFilepath);
+    // Suggest a filename of the form 'consolidated{num}.{extension}', where extension is the extension of the active document
+    let suggestedFilename = generateUnusedFilename(path.extname(documentAbsFilepath), documentDir, 'consolidated');
     // Ask the user for a filename
     const filename = await vscode.window.showInputBox({
         prompt: 'Enter the filename where the consolidated file contents will be saved.',
-        value: 'consolidated' + path.extname(documentAbsFilepath)
+        value: suggestedFilename
     });
     if (filename === undefined) {
         return;
@@ -20,10 +37,16 @@ async function epilogCmd_consolidate(client) {
         vscode.window.showErrorMessage('No filename specified.');
         return;
     }
+    let newDocumentFilepath = documentDir + '/' + filename;
+    // If file exists, ask user if they want to overwrite it
+    if (fs.existsSync(newDocumentFilepath)) {
+        const choice = await vscode.window.showWarningMessage(`File "${newDocumentFilepath}" already exists. Do you want to overwrite it?`, 'Yes', 'No');
+        if (choice !== 'Yes') {
+            return; // User chose not to overwrite
+        }
+    }
     // Resolve the full file content of the active document
     let fullFileContent = (0, resolve_full_file_content_js_1.resolveFullFileContent)(documentAbsFilepath);
-    // Get the directory of the active document
-    let documentDir = path.dirname(documentAbsFilepath);
     // Save the full file content to the filename specified by the user
     fs.writeFileSync(documentDir + '/' + filename, fullFileContent);
 }
